disc-golf: Don't destroy/recreate bot; visual effects
Destroying and recreating the bot causes issues when using deathmatch
scoring.  But not destroying and recreating the bot causes issues when
the game mode starts.  Who knows why.  So, for now, at least, actually
starting the deathmatch mode is not supported.  I guess all scoring
will have to be done manually.  *sigh*

So this version does not destroy/recreate the bot when a disc is
picked up and thrown.  It seems to work.

If the host starts the deathmatch mode, a warning is displayed in the
HUD saying that a restart is required.