disc-golf-skirmish: Several changes
The disc is a Torb rather than a Hammond now, which seems to work
better in several ways.  (Players report that it's much easier to putt
because of the smaller hitbox.)

Attempted several fixes to effects creation and destruction.  Still
not perfect, sometimes effects don't appear with more than 4 players,
so limited to 4 for now.

Dummy bots still aren't destroyed when players leave.  *sigh*

Getting good reviews, though.  Several people really enjoyed it.