disc-golf-skirmish: Many changes, add back Blizzard World, etc.
Still trying to get dummy bots to delete when a player leaves, and
occasionally some disc effects don't get destroyed when they should,
and sometimes tee/hole effects inexplicably disappear when a player
leaves.

Added a reset action the host can use to reset tees/holes and bots.
Not listed in HUD yet, but it helps and allows to avoid restarting the
match.