Fix/Tidy: Effects/text visibility; initial spawn point
After much digging through search results, I found this helpful post
on Reddit.com/r/OverwatchCustomGames:

https://old.reddit.com/r/OverwatchCustomGames/comments/bxl7sa/unintuitive_workshop_mechanics/

It explains that, when a player joins a game, effects that were
created as visible to all players are not automatically made visible
to the new player.  I wish that were documented!

The solution is simple: Create effects that are visible to an array of
players, and when a player joins, update that array, setting it to all
players.  This seems to work!  Such a frustrating problem, but such a
simple solution.

This also includes various tidying, deleting unused things and
reordering some.  This should be a good foundation now.

This also includes using tee position 0 as an initial spawn point on
the map.