lucio-horse: Try to avoid players' turns being skipped
It seems like parts of some rules can be paused and finish executing
at the wrong time.  The Event Inspector sometimes shows that a rule's
conditions became false but it was not aborted because of a wait
condition (but it doesn't give more details than that).  These race
conditions are so tiresome.