lucio-horse: More attempts at fixes
Use a regular, non-subroutine rule for when a follower's turn ends.
Inexplicably, using a subroutine for it seems to cause race
conditions, and a regular rule seems to be more reliable.

Also, disable the timeouts for leader and follower.  Those seem to
come with race conditions too.  Maybe they need to be converted to
non-subr rules also.