winston-chess: Attempted improvements
The race conditions in the Workshop continue to be confounding.  It's
like a game of Jenga in which all the pieces are interconnected, so
pushing one piece in causes another to push out.

These changes sort-of work, but not exactly, and I've had enough of it
for today, so saving for later.