disc-golf: Don't increment stroke counter for free disc
Only increment it when playing a hole.  This avoids counting throws of
a free disc toward a subsequently played hole.  The downside is not
showing the number of strokes when throwing a free disc, but that
would only be useful when testing a self-placed tee/hole, so the
change seems worthwhile.  Changing the code to work in both cases
would be more complicated than it's worth now.